Extract req/resp methods to new request.rb from server.rb
